// src/lib/components/_SourceTable/interface.ts
export type ISourceTableBasePagination = boolean | ITablePagination;
// ...
export interface SourceTableState {
  tableDataSource?: any[];
  searchFormData?: any;
  tableFormData?: any;
  createFormData?: any;
  tableColumnDefs?: IColumnDef[];
  loading?: boolean;
  pagination?: ISourceTableBasePagination;
  saveLoading?: boolean;
  resetLoading?: boolean;
}

export type NoticeType = 'info' | 'success' | 'error' | 'warning' | 'loading';

export interface ISourceTableReturnValue {
  state?: SourceTableState;
  type?: NoticeType;
  message?: string;
  fetch?: boolean;
}

export type ISourceTableActionReturnValue =
  | void
  | boolean
  | SourceTableState
  | ISourceTableReturnValue;

export type ISourceTableOnCreate = (
  event: SourceTableState
) => ISourceTableActionReturnValue | Promise<ISourceTableActionReturnValue>;